Add support for PKCE in the OIDC authenticator
Some IdPs, such as Okta, do not support public clients without PKCE.
This commit adds PKCE to the authorization code flow used by Buildbarn's
OIDC authenticator.

PKCE is an extension to OAuth 2.0 added in RFC 7636. It adds an extra
verification in the form of a hashed nonce to prevent clients from
hijacking the code returned in the regular authorization flow. This
commit also makes the Buildbarn OIDC authentication compatible with the
draft OAuth 2.1 standard, which adds PKCE as a strict requirement for
all clients.

The change is backwards compatible and not configurable. The standard
requires all IdPs to ignore PKCE parameters if they do not implement it,
and the SHA-256 schema is mandatory for IdPs implementing PKCE.

PKCE is designed to solve a problem for native clients that cannot
protect a secret. This is not a problem we have; therefore, this commit
only improves compatibility but not the security of Buildbarn's
authentication flow.